вынести бизнес логику на уровень App
